Adicionando mover para o SD card + Corrigindo atualizacao da Avaliação

+++ app/src/main/java/br/com/redrails/torpedos/models/Mensagem.java
@@ -11,7 +11,7 @@ public class Mensagem {
     private String slug;
     private String autor;
     private Integer data;
-    private Integer avaliacao;
+    private Float avaliacao;
     private Boolean favoritada;
     private Boolean enviada;
     private List<Categoria> categorias = new ArrayList<Categoria>();
@@ -20,7 +20,7 @@ public class Mensagem {
 
     }
 
-    public Mensagem(int id, String texto, boolean favoritada, boolean enviada, String autor, String slug, Integer data, Integer avaliacao) {
+    public Mensagem(int id, String texto, boolean favoritada, boolean enviada, String autor, String slug, Integer data, Float avaliacao) {
         super();
         this.id = id;
         this.texto = texto;
@@ -66,11 +66,11 @@ public class Mensagem {
         this.data = data;
     }
 
-    public Integer getAvaliacao() {
+    public Float getAvaliacao() {
         return avaliacao;
     }
 
-    public void setAvaliacao(Integer avaliacao) {
+    public void setAvaliacao(Float avaliacao) {
         this.avaliacao = avaliacao;
     }
